Fix cube example
