+++ src/lib.rs
@@ -14,9 +14,6 @@ static FS_ENTRY_POINT_SRGB: &str = "fs_main_srgb";
 
 pub type RendererResult<T> = Result<T, RendererError>;
 
-#[cfg(feature = "simple_api_unstable")]
-pub mod simple_api;
-
 #[repr(transparent)]
 #[derive(Debug, Copy, Clone)]
 struct DrawVertPod(DrawVert);
